// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/reportUserMappingAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { ReportUserType } from './reportUserType';
import { ReviewStatus } from './reviewStatus';
import { ReviewerDocumentAC } from './reviewerDocumentAC';

export interface ReportUserMappingAC extends BaseModelAC { 
    reportTitle?: string;
    reportUserType: ReportUserType;
    userType?: string;
    reportId: string;
    userId: string;
    name?: string;
    designation?: string;
    email?: string;
    status: ReviewStatus;
    statusName?: string;
    reviewerDocumentList?: Array<ReviewerDocumentAC>;
}